main.py

import logging
import os
import json
import boto3
from botocore.exceptions import ClientError

from datetime import datetime, timezone
from typing import Annotated

from dotenv import load_dotenv
from livekit import agents
from livekit import api, rtc
from livekit.agents import (Agent, AgentSession,
                            JobProcess, RoomInputOptions,
                            RunContext, function_tool)
from livekit.plugins import deepgram, noise_cancellation, openai, silero
from livekit.agents import get_job_context
from pydantic import Field
from twilio.rest import Client
import re

from user_data import UserData
from recording import start_s3_recording

# --- New imports for DB integration ---
import sqlite3
from db_manager import (
    DB_PATH, create_campaign, add_question, record_call, record_answer
)

load_dotenv()

logger = logging.getLogger("futures_survey_assistant")
logger.setLevel(logging.INFO)
    
RunContext_T = RunContext[UserData]

def get_campaign_from_db(db_path=DB_PATH):
    with sqlite3.connect(db_path) as conn:
        cur = conn.cursor()
        cur.execute("SELECT id, name, description, intro_prompt, purpose_explanation, greeting, closing FROM campaign ORDER BY id DESC LIMIT 1")
        row = cur.fetchone()
        if not row:
            raise Exception("No campaign found in database.")
        return {
            "id": row[0],
            "name": row[1],
            "description": row[2],
            "intro_prompt": row[3],
            "purpose_explanation": row[4],
            "greeting": row[5],
            "closing": row[6],
        }

def get_questions_for_campaign(campaign_id, db_path=DB_PATH):
    with sqlite3.connect(db_path) as conn:
        cur = conn.cursor()
        cur.execute("SELECT id, question_text, question_order FROM question WHERE campaign_id = ? ORDER BY question_order", (campaign_id,))
        return cur.fetchall()

def build_dynamic_prompt_from_db():
    campaign = get_campaign_from_db()
    questions = get_questions_for_campaign(campaign["id"])
    current_time = datetime.now().strftime('%A, %B %d, %Y at %I:%M %p')
    questions_section = ""
    for qid, qtext, qorder in questions:
        questions_section += f"\n{qorder}) Question {qorder}:\n   \"{qtext}\"\n"
    prompt = f"""
{campaign['intro_prompt']}
Current date and time: {current_time}

LANGUAGE POLICY
Detect the participant's first reply.
Do not switch languages once the conversation has started, even if the participant does.
Never use special characters such as %, $, #, or *.

SURVEY FLOW (ask only one question at a time)

1) Briefly explain purpose:
   \"{campaign['purpose_explanation']}\"
{questions_section}
{len(questions) + 3}) Completion check:
   After the recap, call check_survey_complete to ensure all questions were answered.

{len(questions) + 4}) Closing:
   If complete, say:
   \"{campaign['closing']}\"
   Then immediately end the call using the end_call function.

GENERAL GUIDELINES
Ask only one question at a time.
Respond in clear, complete sentences.
If the participant provides unexpected information, politely steer them back to the current question.
Do not provide medical or technical advice; clarify that your role is limited to conducting this survey.
If the participant asks for information outside your scope, respond succinctly that you can only administer the survey.
"""
    return prompt, campaign, questions

class MainAgent(Agent):
    def __init__(self) -> None:
        MAIN_PROMPT, self.campaign, self.questions = build_dynamic_prompt_from_db()
        logger.info("MainAgent initialized with dynamic prompt: %s", MAIN_PROMPT)
        super().__init__(
            instructions=MAIN_PROMPT,
            tools=[set_questionnaire_answer, check_survey_complete],
            tts=openai.TTS(voice="nova"),
        )
    async def on_enter(self) -> None:
        await self.session.say(
            self.campaign["greeting"] or "Hello, welcome to our survey.",
            allow_interruptions=False,
        )

def prewarm(proc: JobProcess):
    proc.userdata["vad"] = silero.VAD.load()

# --- Remove S3 JSON save, use DB instead ---
async def save_userdata_to_db(userdata: UserData, campaign_id: int, call_id: int):
    # Save S3 recording URL if present
    if getattr(userdata, 's3_recording_url', None):
        with sqlite3.connect(DB_PATH) as conn:
            cur = conn.cursor()
            cur.execute("UPDATE call SET s3_recording_url = ? WHERE id = ?", (userdata.s3_recording_url, call_id))
            conn.commit()
            logger.info(f"Updated call {call_id} with S3 recording URL: {userdata.s3_recording_url}")
    elif getattr(userdata, 'recording_id', None):
        # Optionally, if you have a way to build the S3 URL from recording_id, do it here
        pass
    # Save all answers to DB
    for q_num, answer in userdata.questionnaire_answers.items():
        with sqlite3.connect(DB_PATH) as conn:
            cur = conn.cursor()
            cur.execute("SELECT id FROM question WHERE campaign_id = ? AND question_order = ?", (campaign_id, int(q_num)))
            row = cur.fetchone()
            if row:
                question_id = row[0]
                record_answer(call_id, question_id, answer)
                logger.info(f"Saved answer for question {q_num} to DB.")
            else:
                logger.warning(f"Question id not found for campaign {campaign_id}, order {q_num}")
    return True
    
@function_tool    
async def set_questionnaire_answer(
    question_number: Annotated[str, Field(description="The question number (e.g., '1', '2', '3')")],
    answer: Annotated[str, Field(description="The answer")], 
    ctx: RunContext_T
) -> str:
    userdata = ctx.userdata
    userdata.questionnaire_answers[question_number] = answer
    logger.info(f"Question {question_number} answer set: {answer}")
    logger.info(f"All questionnaire answers: {userdata.questionnaire_answers}")
    if len(userdata.questionnaire_answers) == len(userdata.questions):
        return f"Answer for question {question_number} has been saved successfully. Survey complete - ready for finalization: {answer}"
    else:
        return f"Answer for question {question_number} has been saved successfully: {answer}"

@function_tool
async def check_survey_complete(ctx: RunContext_T) -> str:
    userdata = ctx.userdata
    total_questions = len(userdata.questions)
    answered_questions = len(userdata.questionnaire_answers)
    logger.info(f"Survey completion check: {answered_questions}/{total_questions} questions answered")
    if answered_questions == total_questions:
        # Save complete survey to DB
        await save_userdata_to_db(userdata, userdata.campaign["id"], userdata.call_id)
        logger.info("Survey completed - all data saved to DB")
        return f"Survey is complete! All {total_questions} questions have been answered and data has been saved to the database."
    else:
        missing_questions = [str(q[2]) for q in userdata.questions if str(q[2]) not in userdata.questionnaire_answers]
        return f"Survey is not complete. {answered_questions}/{total_questions} questions answered. Missing questions: {missing_questions}"

def extract_phone_from_room_name(room_name: str) -> str:
    pattern = r'call-_(\+\d+)_'
    match = re.search(pattern, room_name)
    if match:
        return match.group(1)
    return None
    
async def entrypoint(ctx: agents.JobContext):
    room = ctx.room
    room_name = room.name
    phone_number = extract_phone_from_room_name(room_name)
    userdata = UserData()
    userdata.customer_phone = phone_number if phone_number else None
    logger.info(f"Room name: {room_name}")
    logger.info(f"Phone number: {phone_number}")
    userdata.agents.update({
        "main_agent": MainAgent(),
    })
    userdata.questions = userdata.agents["main_agent"].questions  # <-- Add this line
    # Start S3 voice recording before recording the call in the DB
    recording_success = await start_s3_recording(room_name, userdata)
    if recording_success:
        logger.info("S3 Recording started successfully")
    else:
        logger.warning("S3 Recording failed, continuing without recording")
        userdata.s3_recording_url = None  # Explicitly set to None if failed
    # Record the call in the DB
    campaign = get_campaign_from_db()
    call_id = record_call(phone_number or "unknown", campaign["id"], s3_recording_url=userdata.s3_recording_url)
    userdata.call_id = call_id
    userdata.campaign = campaign  # Store campaign dict in userdata
    logger.info(f"Call recorded in DB with id: {call_id}")
    await ctx.connect()
    session = AgentSession(
        userdata=userdata,
        stt=deepgram.STT(model="nova-3", language="en-US"),
        llm=openai.LLM(model="gpt-4o-mini"),
        tts=openai.TTS(voice="nova"),
        vad=silero.VAD.load(),
        max_tool_steps=5,
    )
    userdata.session = session
    await session.start(
        agent=userdata.agents["main_agent"],
        room=ctx.room,
        room_input_options=RoomInputOptions(
            noise_cancellation=noise_cancellation.BVC(),
        ),
    )

if __name__ == "__main__": 
    agents.cli.run_app(agents.WorkerOptions(entrypoint_fnc=entrypoint, prewarm_fnc=prewarm, agent_name="alex-telephony-agent"))
